create pyetest_indexd.indexd_server_fixture
